update(grammars): Elect Cobol precise ELS as default
Move Cobol precise ExternalLexStates from the staged cobol_precise_els tag to the default grammar registration, and replace the staged-only regression with a default election test that checks the table, C-recovery gate, and LexModes bounds.

Handle fixed-format trailing comment lines as Cobol trivia, lifting the default 40-file Cobol tier witness from 1/40 to 3/40, and update the tier/election ledgers from staged IV-version 0/40 to default-elected IV-recovery 3/40.
