+++ cgo_harness/tier_scan/tier_classification.tsv
@@ -25,7 +25,7 @@ chatito	IV-unknown	1/5	diagnostic scan 20260621T035630Z-206-inventory-telemetry:
 circom	IV-shape?	19/40	ext=.circom; preliminary
 clojure	IV-perf	unmeasured	diagnostic scan 20260621T035630Z-206-inventory-telemetry: unmeasured (timeout external); scan metadata oom_killed=true, use as diagnostic evidence; previous CLEAN row was stale: in ratchet
 cmake	CLEAN	40/40	in ratchet
-cobol	IV-version	0/40	Wave 4 precise-ELS path no longer times out and matches C on BBANK10P.cbl (Docker wave4-cobol-bbank10p-child-preserve: parityMatch=1/1, trunc=0, diverge=0); generated Cobol gates are clean at 40/40 for both direct C parity (20260708_071502-focus-cobol) and real-corpus parity (20260708T141502Z). Published tier remains IV-version because precise ELS is still staged and the default tier runner reports 0/40 with cRecoveryGateReason=external_scanner_requires_precise_externallexstates (wave4-cobol-tier40-docker: trunc=38, errTree=38, diverge=40); do not ratchet Cobol CLEAN until default election or tier-runner tag handling is resolved.
+cobol	IV-recovery	3/40	Wave 4 default precise ELS is elected: BBANK10P, SCUSTOMP, and SSECUREP match C under the default tier runner (Docker wave4-cobol-tier40-default-els-trailing-comment: parityMatch=3/40, trunc=29, errTree=38, diverge=37, medianRatio=2.70x, aggRatio=4.45x). Generated Cobol gates remain clean at 40/40 for both direct C parity (20260708_071502-focus-cobol) and real-corpus parity (20260708T141502Z), and BBANK10P default ELS witness matches (wave4-cobol-bbank10p-default-els). Remaining IV blockers are 29 no_stacks_alive frontier-death files around copybook/procedure recovery plus accepted C-recovery shape gaps such as SCASHDRV and SCHAR00P where Go materializes finer statements but C wraps the tail as a broad ERROR; do not ratchet Cobol CLEAN until those are resolved.
 comment	IV-perf	unmeasured	diagnostic scan 20260621T035630Z-206-inventory-telemetry: unmeasured (timeout external); scan metadata oom_killed=true, use as diagnostic evidence; previous CLEAN row was stale: full-parse stack cap=2 verified 2026-06-12; parity clean, perf floor pending
 commonlisp	IV-recovery?	unmeasured	ext=.lisp,.cl; faithful C error-cost version competition (see spec)
 cooklang	IV-recovery	0/3	ext=.cook; stackcap disproven (cap 2 not kept); Cooklang-gated C recovery plus local punctuation/fence normalization kept 2026-06-14: D-tier improves 0/3 -> 1/3, trunc 3 -> 0, panics 0. frontmatter_test_recipe now matches; residuals are complex_test_recipe root child-count 51 vs C 42 and test_recipe metadata child-shape (root[2]/root[3] metadata child counts 0 vs C 3/2) after full-span accepted recipe roots.
